requirements.txt
fastapi
uvicorn
jinja2
python-multipart 
python-dotenv
gunicorn
pandas
simpledbf
